refactor(sandbox): split host_check.rs into per-check modules
Issue #28: Split 491-line host_check.rs into checks/ directory with:
- checks/mod.rs — re-exports and Status, Outcome, Report types
- checks/status.rs — Status enum
- checks/outcome.rs — Outcome, Report, Display impl, plural
- checks/run.rs — run() entry point
- checks/linux/mod.rs — Linux runner assembling 7 checks
- checks/linux/*.rs — one file per check (kernel_version, user_namespaces,
  cgroup2, brokkr_slice, seccomp, memory_peak, setgroups)
- host_check.rs — backward-compat re-export for existing worker callers
